chore(modules) Organize database
