+++ src/database/firebase_list_observable.ts
@@ -2,19 +2,8 @@ import {Observable} from 'rxjs/Observable';
 import {Operator} from 'rxjs/Operator';
 import {Subscriber} from 'rxjs/Subscriber';
 import {Subscription} from 'rxjs/Subscription';
-import * as utils from './utils';
-
-export interface FirebaseOperationCases {
-  stringCase: () => firebase.Promise<void>;
-  firebaseCase?: () => firebase.Promise<void>;
-  snapshotCase?: () => firebase.Promise<void>;
-  unwrappedSnapshotCase?: () => firebase.Promise<void>;
-}
-
-export interface AFUnwrappedDataSnapshot {
-  $key: string;
-  $value?: string | number | boolean;
-}
+import * as utils from '../utils/utils';
+import {AFUnwrappedDataSnapshot, FirebaseOperationCases} from '../interfaces';
 
 export type FirebaseOperation = string | firebase.database.Reference | firebase.database.DataSnapshot | AFUnwrappedDataSnapshot;
