+++ src/database/firebase_object_factory.ts
@@ -1,11 +1,16 @@
-import {FirebaseObjectObservable} from './firebase_object_observable';
-import {Observer} from 'rxjs/Observer';
-import 'rxjs/add/operator/mergeMap';
+import { FirebaseObjectObservable } from './index';
+import { Observer } from 'rxjs/Observer';
 import { database } from 'firebase';
-import * as utils from './utils';
-import {Query, observeQuery} from './query_observable';
+import * as utils from '../utils/utils';
+import { Query } from '../interfaces';
+import { observeQuery } from './query_observable';
+import { FirebaseObjectFactoryOpts } from '../interfaces';
+import 'rxjs/add/operator/mergeMap';
 
-export function FirebaseObjectFactory(absoluteUrlOrDbRef: string | firebase.database.Reference, {preserveSnapshot, query}: FirebaseObjectFactoryOpts = {}): FirebaseObjectObservable<any> {
+export function FirebaseObjectFactory (
+  absoluteUrlOrDbRef: string | firebase.database.Reference, 
+  { preserveSnapshot, query }: FirebaseObjectFactoryOpts = {}): FirebaseObjectObservable<any> {
+  
   let ref: firebase.database.Reference;
 
   utils.checkForUrlOrFirebaseRef(absoluteUrlOrDbRef, {
@@ -22,9 +27,4 @@ export function FirebaseObjectFactory(absoluteUrlOrDbRef: string | firebase.data
 
     return () => ref.off();
   }, ref);
-}
-
-export interface FirebaseObjectFactoryOpts {
-  preserveSnapshot?: boolean;
-  query?: Query
 }
